Moved middleware to internal folder.
